+++ src/commonMain/kotlin/com/metrolist/innertubex/models/response/ImageUploadResponse.kt
@@ -1,0 +1,8 @@
+package com.metrolist.innertubex.models.response
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+internal data class ImageUploadResponse(
+    val encryptedBlobId: String,
+)